refactor module structure
renamed .torch to .ml and removed sub-directories for .architectures and .models. Instead, the .ml folder is now flat with all of the modules included.

Also removed empty modules and test files, and moved contents of audio_tools into `audio` module.
