+++ opensoundscape/__init__.py
@@ -3,7 +3,7 @@ __version__ = "0.8.0"
 from . import annotations
 from . import audio_tools
 from . import audio
-from . import audiomoth
+from . import aru
 from . import data_selection
 from . import helpers
 from . import localization
@@ -12,19 +12,20 @@ from . import ribbit
 from . import signal_processing
 from . import spectrogram
 from . import taxa
-from . import torch
+from . import ml
 from . import preprocess
 from . import resources
 from . import wandb
 
 # expose some modules at the top level
-from .torch.models import cnn
+from .ml import cnn
 from .preprocess import preprocessors, actions
 
 # expose some classes at the top level
 from .audio import Audio
 from .spectrogram import Spectrogram
-from .torch.models.cnn import CNN
-from .torch.datasets import AudioFileDataset, AudioSplittingDataset
+from .ml.cnn import CNN, load_model
+from .ml.datasets import AudioFileDataset, AudioSplittingDataset
 from .preprocess.actions import Action
 from .preprocess.preprocessors import SpectrogramPreprocessor
+from .sample import AudioSample
