+++ opensoundscape/ml/__init__.py
@@ -1,0 +1,8 @@
+from . import cam
+from . import cnn_architectures
+from . import cnn
+from . import datasets
+from . import loss
+from . import safe_dataset
+from . import sampling
+from . import utils
