--- opensoundscape/torch/__init__.py
@@ -1,6 +1,0 @@
-from . import grad_cam
-from . import loss
-from . import safe_dataset
-from . import sampling
-from . import architectures
-from . import models
