--- opensoundscape/torch/architectures/__init__.py
@@ -1,2 +1,0 @@
-from . import cnn_architectures
-from . import utils
