--- opensoundscape/torch/grad_cam.py
@@ -1,123 +1,0 @@
-"""
-GradCAM is a method of visualizing the activation of the network on parts
-of an image
-
-# Author:   Kazuto Nakashima
-# URL:      http://kazuto1011.github.io
-# Created:  2017-05-26
-
-"""
-from collections import OrderedDict
-
-import torch
-import torch.nn as nn
-from torch.nn.functional import softmax, adaptive_avg_pool2d
-
-
-class _PropagationBase(object):
-    def __init__(self, model):
-        super(_PropagationBase, self).__init__()
-        self.device = next(model.parameters()).device
-        self.model = model
-        self.image = None
-
-    def _encode_one_hot(self, idx):
-        one_hot = torch.FloatTensor(1, self.preds.size()[-1]).zero_()
-        one_hot[0][idx] = 1.0
-        return one_hot.to(self.device)
-
-    def forward(self, image):
-        self.image = image.requires_grad_()
-        self.model.zero_grad()
-        self.preds = self.model(self.image)
-        self.probs = softmax(self.preds, dim=1)[0]
-        self.prob, self.idx = self.probs.sort(0, True)
-        return self.prob, self.idx
-
-    def backward(self, idx):
-        one_hot = self._encode_one_hot(idx)
-        self.preds.backward(gradient=one_hot, retain_graph=True)
-
-
-class BackPropagation(_PropagationBase):
-    def generate(self):
-        output = self.image.grad.detach().cpu().numpy()
-        return output.transpose(0, 2, 3, 1)[0]
-
-
-class GuidedBackPropagation(BackPropagation):
-    def __init__(self, model):
-        super(GuidedBackPropagation, self).__init__(model)
-
-        def func_b(module, grad_in, grad_out):
-            # Cut off negative gradients
-            if isinstance(module, nn.ReLU):
-                return (torch.clamp(grad_in[0], min=0.0),)
-
-        for module in self.model.named_modules():
-            module[1].register_backward_hook(func_b)
-
-
-class Deconvolution(BackPropagation):
-    def __init__(self, model):
-        super(Deconvolution, self).__init__(model)
-
-        def func_b(module, grad_in, grad_out):
-            # Cut off negative gradients
-            if isinstance(module, nn.ReLU):
-                return (torch.clamp(grad_out[0], min=0.0),)
-
-        for module in self.model.named_modules():
-            module[1].register_backward_hook(func_b)
-
-
-class GradCAM(_PropagationBase):
-    def __init__(self, model):
-        super(GradCAM, self).__init__(model)
-        self.all_fmaps = OrderedDict()
-        self.all_grads = OrderedDict()
-
-        def func_f(module, input, output):
-            self.all_fmaps[id(module)] = output.detach()
-
-        def func_b(module, grad_in, grad_out):
-            self.all_grads[id(module)] = grad_out[0].detach()
-
-        for module in self.model.named_modules():
-            module[1].register_forward_hook(func_f)
-            module[1].register_backward_hook(func_b)
-
-    def _find(self, outputs, target_layer):
-        for key, value in outputs.items():
-            for module in self.model.named_modules():
-                if id(module[1]) == key:
-                    if module[0] == target_layer:
-                        return value
-        raise ValueError("Invalid layer name: {}".format(target_layer))
-
-    def _normalize(self, grads):
-        l2_norm = torch.sqrt(torch.mean(torch.pow(grads, 2))) + 1e-5
-        return grads / l2_norm
-
-    def _compute_grad_weights(self, grads):
-        grads = self._normalize(grads)
-        return adaptive_avg_pool2d(grads, 1)
-
-    def generate(self, target_layer, target_neuron=None):
-        fmaps = self._find(self.all_fmaps, target_layer)
-        grads = self._find(self.all_grads, target_layer)
-        weights = self._compute_grad_weights(grads)
-
-        if target_neuron:
-            target_neuron_one_hot = torch.zeros(512)
-            target_neuron_one_hot[target_neuron] = 1
-            target_neuron_one_hot = target_neuron_one_hot.reshape((1, 512, 1, 1))
-            weights = target_neuron_one_hot.to(self.device)
-
-        gcam = (fmaps[0] * weights[0]).sum(dim=0)
-        gcam = torch.clamp(gcam, min=0.0)
-
-        gcam -= gcam.min()
-        gcam /= gcam.max()
-
-        return gcam.detach().cpu().numpy()
