--- opensoundscape/torch/models/__init__.py
@@ -1,2 +1,0 @@
-from . import cnn
-from . import utils
