--- tests/test_models_utils.py
@@ -1,2 +1,0 @@
-import pytest
-import torch
